+++ netperf_tool/CMakeLists.txt
@@ -36,6 +36,7 @@ ament_python_install_package(${PROJECT_NAME})
 
 pybind11_add_module(netperf_tool_impl SHARED
   src/client_node.cpp
+  src/client_runner.cpp
   src/node_runner.cpp
   src/netperf_tool.cpp
   src/server_node.cpp
